perf(web): lazy-load tauri status panel
